Small Improvement on getElections
Brief Background:
Because some candidates appear in different elections under
different names (sometimes slightly different, sometimes
altogether new, e.g., Liverpool's Richard Pennant ran in 1780
but by 1984 he was The Right Honorable Lord Penrhyn in the poll
book), I added a field in the candidates_elections table --
`running_as` -- to account for name variations of the same
person.

The tweak just made to getElections.php is just to grab the
'running_as' version of the name, if one exists, when a user
opts to include election results in their API call. So it's
basically "if there's a 'running_as' name, use that as the
candidate name, otherwise go with the default name in the
candidates table."